\name{csv_stack}
\alias{csv_stack}
\title{Stack up a directory of csvs or list of data.frames}
\usage{
  csv_stack(csvs = NULL)
}
\arguments{
  \item{csvs}{a directory full of csv files or a list
  data.frames}
}
\value{
  A data.frame with inconsistent columns filled in
}
\description{
  This functions makes data stacks
}

