\name{gen_ctm_data}
\alias{gen_ctm_data}
\title{output .dat document term format for ctm algorithm}
\usage{
  gen_ctm_data(corpus, file = "ctm.dat")
}
\arguments{
  \item{corpus}{a corpus / character vector of text
  documents}
}
\description{
  output .dat document term format for ctm algorithm
}
\examples{
#gen_ctm_data(corpus, "ctm.dat")
}

